+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/HttpTableLookupFunction.java
@@ -19,6 +19,8 @@ import org.apache.flink.table.data.binary.BinaryStringData;
 import org.apache.flink.table.functions.FunctionContext;
 import org.apache.flink.table.functions.TableFunction;
 
+import com.getindata.connectors.http.LookupArg;
+import com.getindata.connectors.http.LookupQueryCreator;
 import com.getindata.connectors.http.internal.PollingClient;
 import com.getindata.connectors.http.internal.PollingClientFactory;
 
@@ -28,6 +30,8 @@ public class HttpTableLookupFunction extends TableFunction<RowData> {
     private final PollingClientFactory<RowData> pollingClientFactory;
 
     private final DeserializationSchema<RowData> schemaDecoder;
+
+    private final LookupQueryCreator lookupQueryCreator;
 
     @Getter
     private final ColumnData columnData;
@@ -44,19 +48,22 @@ public class HttpTableLookupFunction extends TableFunction<RowData> {
         PollingClientFactory<RowData> pollingClientFactory,
         DeserializationSchema<RowData> schemaDecoder,
         ColumnData columnData,
-        HttpLookupConfig options) {
+        HttpLookupConfig options,
+        LookupQueryCreator lookupQueryCreator) {
 
         this.pollingClientFactory = pollingClientFactory;
         this.schemaDecoder = schemaDecoder;
         this.columnData = columnData;
         this.options = options;
+        this.lookupQueryCreator = lookupQueryCreator;
     }
 
     @Override
     public void open(FunctionContext context) throws Exception {
         super.open(context);
         this.localHttpCallCounter = new AtomicInteger(0);
-        this.client = pollingClientFactory.createPollClient(options, schemaDecoder);
+        this.client = pollingClientFactory
+            .createPollClient(options, schemaDecoder, lookupQueryCreator);
 
         context
             .getMetricGroup()