+++ src/test/java/com/getindata/connectors/http/internal/table/lookup/HttpRowDataLookupFunctionTest.java
@@ -20,9 +20,12 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.getindata.connectors.http.LookupArg;
+import com.getindata.connectors.http.LookupQueryCreator;
 import com.getindata.connectors.http.internal.PollingClient;
 import com.getindata.connectors.http.internal.PollingClientFactory;
 import com.getindata.connectors.http.internal.table.lookup.HttpTableLookupFunction.ColumnData;
+import com.getindata.connectors.http.internal.table.lookup.querycreators.GenericGetQueryCreator;
 
 @ExtendWith(MockitoExtension.class)
 class HttpRowDataLookupFunctionTest {
@@ -50,6 +53,7 @@ class HttpRowDataLookupFunctionTest {
     @BeforeEach
     void setUp() throws Exception {
         HttpLookupConfig lookupConfig = HttpLookupConfig.builder().build();
+        LookupQueryCreator lookupQueryCreator = new GenericGetQueryCreator();
 
         ColumnData columnData =
             ColumnData.builder()
@@ -57,13 +61,15 @@ class HttpRowDataLookupFunctionTest {
                 .build();
 
         when(functionContext.getMetricGroup()).thenReturn(mock(MetricGroup.class));
-        when(pollingClientFactory.createPollClient(lookupConfig, schemaDecoder)).thenReturn(client);
+        when(pollingClientFactory.createPollClient(lookupConfig, schemaDecoder, lookupQueryCreator))
+            .thenReturn(client);
 
         lookupFunction = HttpTableLookupFunction.builder()
             .pollingClientFactory(pollingClientFactory)
             .schemaDecoder(schemaDecoder)
             .columnData(columnData)
             .options(lookupConfig)
+            .lookupQueryCreator(lookupQueryCreator)
             .build();
 
         lookupFunction.open(functionContext);