+++ src/test/java/com/getindata/connectors/http/internal/utils/uri/URIBuilderTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.getindata.connectors.http.internal.table.lookup.LookupArg;
+import com.getindata.connectors.http.LookupArg;
 
 class URIBuilderTest {